some changes

+++ netbox-powerdns-sync.py
@@ -4,17 +4,22 @@ import powerdns
 import pynetbox
 import re
 import ipaddress
+import os
 
 from config import NB_URL, NB_TOKEN, PDNS_API_URL, PDNS_KEY
-from config import FORWARD_ZONES, REVERSE_ZONES
+from config import FORWARD_ZONES, REVERSE_ZONES, REQUESTS_CA_BUNDLE
+from config import DEBUG
 
+os.environ['REQUESTS_CA_BUNDLE'] = REQUESTS_CA_BUNDLE
 nb = pynetbox.api(NB_URL, token=NB_TOKEN)
+#nb.http_session.verify = False
 
 pdns_api_client = powerdns.PDNSApiClient(api_endpoint=PDNS_API_URL, api_key=PDNS_KEY)
 pdns = powerdns.PDNSEndpoint(pdns_api_client).servers[0]
 
 host_ips = []
 record_ips = []
+record_wo_comment_ips = []
 for forward_zone in FORWARD_ZONES:
     forward_zone_canonical = forward_zone+"."
 
@@ -48,6 +53,12 @@ for forward_zone in FORWARD_ZONES:
                                        record["type"],
                                        ip["content"],
                                        forward_zone_canonical))
+        else:
+            for ip in record["records"]:
+                record_wo_comment_ips.append((record["name"],
+                                    record["type"],
+                                    ip["content"],
+                                    forward_zone_canonical))
 
 for reverse_zone in REVERSE_ZONES:
     # get IPs within the prefix from NetBox
@@ -82,16 +93,30 @@ for reverse_zone in REVERSE_ZONES:
 
 # create set with tupels that have to be created
 # tupels from NetBox without tupels that already exists in PowerDNS
-to_create = set(host_ips)-set(record_ips)
+to_create = set(host_ips)-set(record_ips)-set(record_wo_comment_ips)
+
+# create set with tupels that have to be updated
+# tupels from NetBox that already exists in PowerDNS but dont have comment
+to_update = set(host_ips)&set(record_wo_comment_ips)
 
 # create set with tupels that have to be deleted
 # tupels from PowerDNS without tupels that are documented in NetBox
 to_delete = set(record_ips)-set(host_ips)
 
+# create set with tupels that are missing
+# tupels from PowerDNS that are not documented in NetBox
+missing = set(record_wo_comment_ips)-(set(record_ips)-set(host_ips))
+
 print("----")
 
 print(len(to_create), "records to create:")
 for record in to_create:
+    print(record[0])
+
+print("----")
+
+print(len(to_update), "records to update:")
+for record in to_update:
     print(record[0])
 
 print("----")
@@ -102,26 +127,47 @@ for record in to_delete:
 
 print("----")
 
+print(len(missing), "missing records:")
+for record in missing:
+    print(record[0])
+
+print("----")
+
 for record in to_create:
     print("Creating", record)
-    zone = pdns.get_zone(record[3])
-    zone.create_records([
-                        powerdns.RRSet(record[0],
-                                       record[1],
-                                       [(record[2], False)],
-                                       comments=[powerdns.Comment("NetBox")])
-                        ])
+    if not DEBUG:
+        zone = pdns.get_zone(record[3])
+        zone.create_records([
+                            powerdns.RRSet(record[0],
+                                        record[1],
+                                        [(record[2], False)],
+                                        comments=[powerdns.Comment("NetBox")])
+                            ])
+
+print("----")
+
+for record in to_update:
+    print("Updating", record)
+    if not DEBUG:
+        zone = pdns.get_zone(record[3])
+        zone.create_records([
+                            powerdns.RRSet(record[0],
+                                        record[1],
+                                        [(record[2], False)],
+                                        comments=[powerdns.Comment("NetBox")])
+                            ])
 
 print("----")
 
 for record in to_delete:
     print("Deleting", record)
-    zone = pdns.get_zone(record[3])
-    zone.delete_records([
-                        powerdns.RRSet(record[0],
-                                       record[1],
-                                       [(record[2], False)],
-                                       comments=[powerdns.Comment("NetBox")])
-                        ])
+    if not DEBUG:
+        zone = pdns.get_zone(record[3])
+        zone.delete_records([
+                            powerdns.RRSet(record[0],
+                                        record[1],
+                                        [(record[2], False)],
+                                        comments=[powerdns.Comment("NetBox")])
+                            ])
 
 print("----")
